Two changes: 1. Added cross_validate_linear_svm which recieves the following arguments:
train_feature: the training data (numpy.ndarray)
train_labels: the trainig labeles 1,2,...  (numpy.ndarray)
param_name: the parameters that you wish to tune (default to "C")
param_range: range for values of the parameter to tune (defaults to
np.logspace(-2, 2, 10))
cv: number of cross validation folds (defaults to 5)
scoring: type of scoring metric to use in the cross validation (defaults
to "accuracy")
n_jobs: number of job to run concurrently.
random_state: random seed to use in randomizing the order of the data in
the different folds

return values: two arrays with  train_scores and val_scores

2. Added a test function test_read_and_validate which tests both the
readdata() function and the by reading the HAPT Data Set and tuning a
linear SVM on the trainig data.